decorators.py split to validation, utils, patterns
